Proposed update to remove RAII problems
I propose the removal of the use of `std::exit`. It's usage breaks RAII which is not mentioned in the documentation.

As for the addition of the exception, during parsing we already require that the user catches several exceptions thus the whole `try/catch` harness is already there for a proper program. Adding a catch clause for the new `NormalProgramTermination` is a bit bulky but doesn't break RAII and is explicit on that the program should terminate in normal usage.